adde: note push notes

+++ src/note/note.service.ts
@@ -3,17 +3,22 @@ import {
   ForbiddenException,
   Injectable,
   NotFoundException,
-  Type,
 } from '@nestjs/common';
 import { CreateNoteDto } from './dto/create-note.dto';
 import { UpdateNoteDto } from './dto/update-note.dto';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model, Types } from 'mongoose';
 import { Note } from './schema/note.schema';
+import { AuthService } from 'src/auth/auth.service';
+import { UserService } from 'src/user/user.service';
 
 @Injectable()
 export class NoteService {
-  constructor(@InjectModel(Note.name) private noteModel: Model<Note>) {}
+  constructor(
+    @InjectModel(Note.name) private noteModel: Model<Note>,
+    private userService: UserService,
+  ) {}
+
 
   async create(userId: Types.ObjectId | string, createNoteDto: CreateNoteDto) {
     if (!userId) {
@@ -25,6 +30,12 @@ export class NoteService {
         ...createNoteDto,
         userId: userId,
       });
+      const noteOwner = await this.userService.getById(userId);
+      console.log(noteOwner, 'noteOwner');
+      console.log(newNote, 'newNote');
+      if (!noteOwner) throw new NotFoundException('User not found');
+      noteOwner.notes.push(newNote._id);
+      await noteOwner.save();
       return newNote;
     } catch (e) {
       console.log(e);
@@ -32,18 +43,34 @@ export class NoteService {
     }
   }
 
-  async findAll(userId: Types.ObjectId | string) {
+  async findAll(userId: Types.ObjectId | string, isArchived?: boolean) {
     if (!userId) {
       throw new ForbiddenException('Permission denied');
     }
     try {
       const userObjectId = new Types.ObjectId(userId);
-      // console.log(userObjectId, "userObjectId")
-      // console.log(userId, "userId")
-      const usersAllNotes = await this.noteModel
-        .find({ userId: userObjectId })
+      const filter: any = { userId: userObjectId };
+      if (typeof isArchived === 'boolean') {
+        filter.isArchived = isArchived;
+      }
+      const usersNotes = await this.noteModel.find(filter).exec();
+      return usersNotes;
+    } catch (e) {
+      console.error('Error in findAll:', e);
+      throw e;
+    }
+  }
+
+  async findAllArchived(userId: Types.ObjectId | string) {
+    if (!userId) {
+      throw new ForbiddenException('Permission denied');
+    }
+    try {
+      const userObjectId = new Types.ObjectId(userId);
+      const archivedNotes = await this.noteModel
+        .find({ userId: userObjectId, isArchived: true })
         .exec();
-      return usersAllNotes;
+      return archivedNotes;
     } catch (e) {
       console.log(e);
       throw e;
@@ -70,12 +97,22 @@ export class NoteService {
     }
   }
 
+
   async remove(userId: Types.ObjectId | string, id: string) {
     if (!userId) {
       throw new ForbiddenException('Permission denied');
     }
     try {
       if (!id) throw new BadRequestException('Note ID is required');
+      const noteOwner = await this.userService.getById(userId);
+      if (!noteOwner) throw new NotFoundException('User not found');
+      const index = noteOwner.notes.findIndex(
+        (noteId) => noteId.toString() === id.toString(),
+      );
+      if (index !== -1) {
+        noteOwner.notes.splice(index, 1);
+        await noteOwner.save();
+      }
       await this.noteModel.findByIdAndDelete(id);
 
       return 'Note deleted successfully';
@@ -85,11 +122,54 @@ export class NoteService {
     }
   }
 
-  // findOne(id: number) {
-  //   return `This action returns a #${id} note`;
-  // }
+  async update(
+    userId: Types.ObjectId | string,
+    id: string,
+    updateNoteDto: UpdateNoteDto,
+  ) {
+    if (!userId) {
+      throw new ForbiddenException('Permission denied');
+    }
+    try {
+      if (!id || !updateNoteDto) {
+        throw new BadRequestException('Missing id or update data');
+      }
+      const updatedNote = await this.noteModel.findByIdAndUpdate(
+        id,
+        updateNoteDto,
+        { new: true },
+      );
+      if (!updatedNote) {
+        throw new NotFoundException(`Note with ID ${id} not found`);
+      }
+      const message =
+        updatedNote.isArchived === true
+          ? 'Note successfully archived'
+          : 'Note successfully updated and restored';
+      return message;
+    } catch (e) {
+      console.log(e);
+      throw e;
+    }
+  }
 
-  update(id: number, updateNoteDto: UpdateNoteDto) {
-    return `This action updates a #${id} note`;
+  async notesFilteredByTag(userId: Types.ObjectId | string, tag: string) {
+    if (!userId) {
+      throw new ForbiddenException('Permission denied');
+    }
+    try {
+      if (!tag) throw new BadRequestException('Tag is required');
+      const filteredByTag = await this.noteModel.find({
+        userId,
+        tags: { $in: [tag] },
+      });
+      return {
+        filteredByTag,
+        tag,
+      };
+    } catch (e) {
+      console.log(e);
+      throw e;
+    }
   }
 }
